backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
python-multipart==0.0.9
docxtpl==0.16.7
docx2pdf==0.1.8
openpyxl==3.1.5
python-dotenv==1.0.1
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.9.2
num2words